refactoring Metadata-selector files to be in detail folder, Removing IMetadataSummary.ts
