--- ui/src/app/metadata/IMetadataSummary.ts
@@ -1,15 +1,0 @@
-import { ApiMetadata } from "./IMetadata"
-
-export interface IMetadataSummary {
-  count?: number
-  metadata?: ApiMetadata
-}
-
-export class ApiMetadataSummary implements IMetadataSummary {
-  count?: number
-  metadata?: ApiMetadata
-  constructor({count, metadata}: IMetadataSummary)  {
-    this.count = count
-    this.metadata = metadata
-  }
-}
