+++ ui/src/app/metadata/detail/metadata-selector/metadata-selector.component.spec.ts
